requirements/base.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
mysqlclient==2.2.0
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
Pillow==10.1.0
qrcode==7.4.2
reportlab==4.0.7
cloudinary==1.36.0
django-cloudinary-storage==0.3.0
razorpay==1.4.1
python-decouple==3.8
django-filter==23.3
drf-spectacular==0.26.5
